bookmarket

+++ spring/BookClub/src/main/java/com/stack_java/bookclub/controllers/MainController.java
@@ -7,11 +7,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 
 import com.stack_java.bookclub.models.Book;
 import com.stack_java.bookclub.models.User;
@@ -24,15 +25,18 @@ public class MainController {
 
 	@Autowired
 	private UserService userService;
-	
+
 	@Autowired
 	private BooksService bookService;
 
 //	LOGIN & REGISTRATION
 	@GetMapping("/")
-	public String index(Model model) {
+	public String index(Model model, HttpSession session) {
 		model.addAttribute("newUser", new User());
 		model.addAttribute("newLogin", new UserLogin());
+		if (session.getAttribute("userId") != null) {
+			return "redirect:/dashboard";
+		}
 		return "index.jsp";
 	}
 
@@ -53,7 +57,7 @@ public class MainController {
 	public String register(@Valid @ModelAttribute("newLogin") UserLogin newLogin, BindingResult result, Model model,
 			HttpSession session) {
 		User user = userService.login(newLogin, result);
-		if(result.hasErrors()) {
+		if (result.hasErrors()) {
 			model.addAttribute("newUser", new User());
 			return "index.jsp";
 		}
@@ -61,77 +65,123 @@ public class MainController {
 		session.setAttribute("userName", user.getName());
 		return "redirect:/dashboard";
 	}
-	
+
 //	Logout
 	@GetMapping("/logout")
 	public String logout(HttpSession session) {
 		session.invalidate();
 		return "redirect:/";
 	}
-	
-	
+
 //	Dashboard
 	@GetMapping("/dashboard")
 	public String dashboard(HttpSession session, Model model) {
 		model.addAttribute("userId", session.getAttribute("userId"));
-		model.addAttribute("userName" ,session.getAttribute("userName"));
+		model.addAttribute("userName", session.getAttribute("userName"));
 		model.addAttribute("allBooks", bookService.allBooks());
-		if(model.getAttribute("userId") == null) {
+		if (model.getAttribute("userId") == null) {
 			return "redirect:/";
 		}
 		return "dashboard.jsp";
 	}
-	
+
 //	Create a book
 	@GetMapping("/book/create")
 	public String book(HttpSession session, Model model) {
 		model.addAttribute("userId", session.getAttribute("userId"));
-		model.addAttribute("userName" ,session.getAttribute("userName"));
-		if(model.getAttribute("userId") == null) {
+		model.addAttribute("userName", session.getAttribute("userName"));
+		if (model.getAttribute("userId") == null) {
 			return "redirect:/";
 		}
 		model.addAttribute("newBook", new Book());
 		return "createBook.jsp";
 	}
-	
+
 	@PostMapping("/book/create/success")
-	public String createBook(@Valid @ModelAttribute("newBook") Book book, BindingResult result, Model model) {
-		if(result.hasErrors()) {
+	public String createBook(@Valid @ModelAttribute("newBook") Book book, BindingResult result, HttpSession session) {
+		if (session.getAttribute("userId") == null) {
+			return "redirect:/";
+		}
+		if (result.hasErrors()) {
 			return "createBook.jsp";
 		}
 		bookService.createBook(book);
 		return "redirect:/dashboard";
 	}
-	
+
 //	Show One Book
 	@GetMapping("/book/show/{id}")
-	public String showBook(@PathVariable("id") Long id, Model model) {
+	public String showBook(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null) {
+			return "redirect:/";
+		}
 		model.addAttribute("oneBook", bookService.getBook(id));
 		return "oneBook.jsp";
 	}
-	
+
 //	edit book
 	@GetMapping("/book/edit/{id}")
-	public String editBookShow(@PathVariable("id") Long id, Model model) {
+	public String editBookShow(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if (session.getAttribute("userId") != id) {
+			return "redirect:/";
+		}
 		model.addAttribute("oneBook", bookService.getBook(id));
 		return "editBook.jsp";
 	}
 	
-	@PostMapping("/book/edit/success")
-	public String editBook(@Valid @ModelAttribute("oneBook") Book book, BindingResult result, Model model) {
-		if(result.hasErrors()) {
+//	Make sure to use PutMapping for edits
+	@PutMapping("/book/edit/success")
+	public String editBook(@Valid @ModelAttribute("oneBook") Book book, BindingResult result, Model model,
+			HttpSession session) {
+		if (session.getAttribute("userId") != book.getUser().getId()) {
+			return "redirect:/dashboard";
+		}
+		if (result.hasErrors()) {
 			return "editBook.jsp";
 		}
 		bookService.createBook(book);
 		return "redirect:/dashboard";
 	}
-	
+
 //	delete book
-	@RequestMapping("/books/delete/{id}")
-//	 what is going on here?????????????????? why cant i use DeleteMapping and why can i use request?
-    public String deleteBook(@PathVariable("id") Long id) {
-        bookService.deleteBook(id);
-        return "redirect:/dashboard";
-    }
+	@DeleteMapping("/book/delete/{id}")
+	public String destroy(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if (!session.getAttribute("userId").equals(bookService.getBook(id).getUser().getId())) {
+			return "redirect:/dashboard";
+		}
+		bookService.deleteBook(id);
+		return "redirect:/dashboard";
+	}
 	
+//	wrong way
+//	@RequestMapping("/book/delete/{id}")
+//    public String deleteBook(@PathVariable("id") Long id, HttpSession session) {
+//		if (!session.getAttribute("userId").equals(bookService.getBook(id).getUser().getId())) {
+//		return "redirect:/dashboard";
+//	}
+//        bookService.deleteBook(id);
+//        return "redirect:/dashboard";
+//    }
+	
+//	book broker dashboard
+	@GetMapping("/bookmarket")
+	public String bookMarket(HttpSession session, Model model) {
+		model.addAttribute("userName", session.getAttribute("userName"));
+		model.addAttribute("userId", session.getAttribute("userId"));
+		if (session.getAttribute("userId") == null) {
+			return "redirect:/";
+		}
+		model.addAttribute("allBooks", bookService.allBooks());
+		model.addAttribute("oneUser", userService.oneUser(session.getAttribute("userId")));
+		return "bookMarket.jsp";
+	}
+	
+	@GetMapping("/borrow/add/{id}")
+	public String borrow(@PathVariable("id") Long id, HttpSession session) {
+		if (session.getAttribute("userId") == null) {
+			return "redirect:/";
+		}
+		
+		return "redirect:/bookmarket";
+	}
 }
